refactor: extract board state module
